requirements.txt
# Web framework
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9

# NLP
spacy==3.7.5
sentence-transformers==3.0.1

# LLM
groq==0.9.0

# Document parsing
pymupdf==1.24.9
python-docx==1.1.2

# PDF generation
reportlab==4.2.2

# Utilities
python-dotenv==1.0.1
numpy==1.26.4
httpx==0.27.0
pydantic==2.7.4

# Streamlit
streamlit==1.37.0
requests==2.32.3